refactor bluebird tests to avoid timeout
