+++ test/versioned/bluebird/package.json
@@ -24,7 +24,41 @@
         "bluebird": ">=3.0.0"
       },
       "files": [
-        "methods.test.js"
+        "baseline.test.js",
+        "promise-all.test.js",
+        "promise-all-settled.test.js",
+        "promise-any.test.js",
+        "promise-bind.test.js",
+        "promise-call.test.js",
+        "promise-catch-return.test.js",
+        "promise-catch-throw.test.js",
+        "promise-coroutine.test.js",
+        "promise-constructor.test.js",
+        "promise-delay.test.js",
+        "promise-each.test.js",
+        "promise-error.test.js",
+        "promise-filter.test.js",
+        "promise-get.test.js",
+        "promise-get-new-library-copy.test.js",
+        "promise-is.test.js",
+        "promise-join.test.js",
+        "promise-map.test.js",
+        "promise-map-series.test.js",
+        "promise-method.test.js",
+        "promise-no-conflict.test.js",
+        "promise-promisify.test.js",
+        "promise-props.test.js",
+        "promise-race.test.js",
+        "promise-reduce.test.js",
+        "promise-reflect.test.js",
+        "promise-return.test.js",
+        "promise-some.test.js",
+        "promise-spread.test.js",
+        "promise-tap.test.js",
+        "promise-tap-catch.test.js",
+        "promise-then.test.js",
+        "promise-then-return.test.js",
+        "promise-timeout.test.js"
       ]
     }
   ],
